backend/requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.0
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==5.1
django-stubs==5.0.4
django-stubs-ext==5.0.4
frozenlist==1.4.1
future==1.0.0
idna==3.7
line-bot-sdk==3.11.1
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
types-PyYAML==6.0.12.20240808
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
